Firefox: save user.js to profile
